Take a crack at reorganizing in a way that better reflects how a larger project might be laid out
