+++ README.md
@@ -3,6 +3,8 @@ The secondary purpose of this project is to implement an interpreter for the
 [Brainfuck programming language](https://en.wikipedia.org/wiki/Brainfuck) with
 a few extensions. The primary purpose is to learn Rust.
 
+
+
 ## Usage
 The interpreter can be compiled and run:
 ```
@@ -14,12 +16,16 @@ $ cargo build --release
 $ ./target/release/bf <program>
 ```
 
+
+
 ## Language Extensions
 This implementation follows the
 [Wikipedia standard](https://en.wikipedia.org/wiki/Brainfuck#Commands) with
 some additions:
 - `?`: Dump the internal program execution state to stderr
 - `!`: Breakpoint to enter into a Brainfuck REPL
+
+
 
 ## Goals
 - No `panic!`
@@ -32,6 +38,8 @@ some additions:
     - Proper exit code setting
     - Expected options like `--help`
 - Reasonable performance and resource usage (nothing dumb)
+
+
 
 ## Example Programs
 1. `hello world`:
@@ -47,8 +55,8 @@ some additions:
 
 3. `Hello World!`:
 ```
-'>++++++++[-<+++++++++>]<.>>+>-[+]++>++>+++[>[->+++<<+++>]<<]>-----.>->+++..+++
-.>-.<<+[>[+>+]>>]<--------------.>>.+++.------.--------.>+.>+.'
+>++++++++[-<+++++++++>]<.>>+>-[+]++>++>+++[>[->+++<<+++>]<<]>-----.>->+++..+++.
+>-.<<+[>[+>+]>>]<--------------.>>.+++.------.--------.>+.>+.
 ```
 
 4. `Hello, World!`:
@@ -60,4 +68,13 @@ some additions:
 ```
 ,[.[-],]
 ```
+
     - Example: `$ ./bf ',[.[-],]' < README.md`
+
+## TODOs
+- [ ] Accept and ignore non-command characters instead of failing (comments)
+- [ ] Integrate `readline` for REPL input
+- [ ] Write docstrings
+- [ ] Support running from file with `-f` and `--file`
+- [ ] Add `-h`/`--help` usage flag
+- [ ] Add `-v`/`--verbose` flag to print extra information (like exit message)
