+++ src/main.rs
@@ -1,6 +1,8 @@
 use std::env;
 
-mod bf;
+mod token;
+mod interpreter;
+mod repl;
 
 fn main() {
     let err_str: &'static str = "usage: bf <program>";
@@ -8,13 +10,13 @@ fn main() {
     // assert_eq!(args.len(), 2);
     std::process::exit(match &args[..] {
         [_, program] => {
-            let state = bf::run_interpreter(program);
+            let state = interpreter::run(program);
             match state.status {
-                bf::ExecutionStatus::Terminated => {
+                interpreter::ExecutionStatus::Terminated => {
                     eprintln!("bf interpreter terminated: 0");
                     0
                 },
-                bf::ExecutionStatus::Error(err) => {
+                interpreter::ExecutionStatus::Error(err) => {
                     eprintln!("bf interpreter exited with error: {}", err);
                     1
                 },
